Switch to sphinx-book-theme, add header templates
For now, these are kept in _templates for docs_src and manual.
We should change these to use our own Sphinx theme that can take in
a new parameter (like our previous commit did with pydata-sphinx-theme)
to set the URLs to show.

There's a bit of extra CSS neded as well, to make sure that the
logo is correctly shrunk when in the navbar (rather than when in the
sidebar). If these are left out, we get a huge logo across the page.

Also we need to decide what to do with the logo(s) on the top left.